PR feedback + fix test

+++ node/pkg/watchers/near/watcher_test.go
@@ -589,7 +589,7 @@ func TestSuccessValueToInt(t *testing.T) {
 
 	type test struct {
 		input  string
-		output int
+		output uint64
 	}
 
 	testsPositive := []test{
